requirements.txt
# FastAPI and dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.2

# File processing
pdfplumber>=0.10.0
python-docx>=1.1.0
Pillow>=10.0.0

# PDF export
weasyprint>=60.0
reportlab>=4.0.0

# Existing dependencies
langchain>=0.2.0
langchain-core>=0.2.0
langchain-anthropic>=0.1.1
langchain-openai>=0.1.0
langchain-google-genai>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0